Update base image to `golang:1.22-alpine3.18` (#49)
This PR updates the base image as part of the Golang 1.22 update.

N.B: We are still building the images for `1.19`. We'll still need these while we update all our services. The `1.19` image will now use `alpine3.16` as to avoid over complicate some code that will be removed soon (1.19 will be phased out pretty quickly)